Split TranslationServer into its own file
