+++ core/string/translation_server.compat.inc
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  translation.h                                                         */
+/*  translation_server.compat.inc                                         */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -28,50 +28,11 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef TRANSLATION_H
-#define TRANSLATION_H
-
-#include "core/io/resource.h"
-#include "core/object/gdvirtual.gen.inc"
-
-class Translation : public Resource {
-	GDCLASS(Translation, Resource);
-	OBJ_SAVE_TYPE(Translation);
-	RES_BASE_EXTENSION("translation");
-
-	String locale = "en";
-	HashMap<StringName, StringName> translation_map;
-
-	virtual Vector<String> _get_message_list() const;
-	virtual Dictionary _get_messages() const;
-	virtual void _set_messages(const Dictionary &p_messages);
-
-	void _notify_translation_changed_if_applies();
-
-protected:
-	static void _bind_methods();
-
 #ifndef DISABLE_DEPRECATED
-	static void _bind_compatibility_methods();
+
+void TranslationServer::_bind_compatibility_methods() {
+	ClassDB::bind_compatibility_method(D_METHOD("translate", "message", "context"), &TranslationServer::translate, DEFVAL(""));
+	ClassDB::bind_compatibility_method(D_METHOD("translate_plural", "message", "plural_message", "n", "context"), &TranslationServer::translate_plural, DEFVAL(""));
+}
+
 #endif
-
-	GDVIRTUAL2RC(StringName, _get_message, StringName, StringName);
-	GDVIRTUAL4RC(StringName, _get_plural_message, StringName, StringName, int, StringName);
-
-public:
-	void set_locale(const String &p_locale);
-	_FORCE_INLINE_ String get_locale() const { return locale; }
-
-	virtual void add_message(const StringName &p_src_text, const StringName &p_xlated_text, const StringName &p_context = "");
-	virtual void add_plural_message(const StringName &p_src_text, const Vector<String> &p_plural_xlated_texts, const StringName &p_context = "");
-	virtual StringName get_message(const StringName &p_src_text, const StringName &p_context = "") const; //overridable for other implementations
-	virtual StringName get_plural_message(const StringName &p_src_text, const StringName &p_plural_text, int p_n, const StringName &p_context = "") const;
-	virtual void erase_message(const StringName &p_src_text, const StringName &p_context = "");
-	virtual void get_message_list(List<StringName> *r_messages) const;
-	virtual int get_message_count() const;
-	virtual Vector<String> get_translated_message_list() const;
-
-	Translation() {}
-};
-
-#endif // TRANSLATION_H
