+++ core/string/ustring.cpp
@@ -36,7 +36,7 @@
 #include "core/os/memory.h"
 #include "core/string/print_string.h"
 #include "core/string/string_name.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 #include "core/string/ucaps.h"
 #include "core/variant/variant.h"
 #include "core/version_generated.gen.h"
