+++ editor/editor_node.cpp
@@ -42,7 +42,7 @@
 #include "core/os/os.h"
 #include "core/os/time.h"
 #include "core/string/print_string.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 #include "core/version.h"
 #include "editor/editor_string_names.h"
 #include "main/main.h"
