+++ editor/import/resource_importer_csv_translation.cpp
@@ -33,7 +33,7 @@
 #include "core/io/file_access.h"
 #include "core/io/resource_saver.h"
 #include "core/string/optimized_translation.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 
 String ResourceImporterCSVTranslation::get_importer_name() const {
 	return "csv_translation";
