+++ editor/localization_editor.cpp
@@ -31,7 +31,7 @@
 #include "localization_editor.h"
 
 #include "core/config/project_settings.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 #include "editor/editor_translation_parser.h"
 #include "editor/editor_undo_redo_manager.h"
 #include "editor/filesystem_dock.h"
