+++ editor/plugins/font_config_plugin.cpp
@@ -30,6 +30,7 @@
 
 #include "font_config_plugin.h"
 
+#include "core/string/translation_server.h"
 #include "editor/editor_settings.h"
 #include "editor/import/dynamic_font_import_settings.h"
 #include "editor/themes/editor_scale.h"
