+++ editor/project_manager/quick_settings_dialog.cpp
@@ -31,7 +31,7 @@
 #include "quick_settings_dialog.h"
 
 #include "core/config/project_settings.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 #include "editor/editor_settings.h"
 #include "editor/editor_string_names.h"
 #include "editor/themes/editor_scale.h"
