+++ modules/text_server_adv/text_server_adv.cpp
@@ -51,7 +51,7 @@ using namespace godot;
 #include "core/error/error_macros.h"
 #include "core/object/worker_thread_pool.h"
 #include "core/string/print_string.h"
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 #include "scene/resources/image_texture.h"
 
 #include "modules/modules_enabled.gen.h" // For freetype, msdfgen, svg.
