+++ tests/core/string/test_translation.h
@@ -34,6 +34,7 @@
 #include "core/string/optimized_translation.h"
 #include "core/string/translation.h"
 #include "core/string/translation_po.h"
+#include "core/string/translation_server.h"
 
 #ifdef TOOLS_ENABLED
 #include "editor/import/resource_importer_csv_translation.h"
