+++ tests/core/string/test_translation_server.h
@@ -31,7 +31,7 @@
 #ifndef TEST_TRANSLATION_SERVER_H
 #define TEST_TRANSLATION_SERVER_H
 
-#include "core/string/translation.h"
+#include "core/string/translation_server.h"
 
 #include "tests/test_macros.h"
